+++ modules/delegates/api/fetchMKRWeightHistory.ts
@@ -1,5 +1,5 @@
 import { SupportedNetworks } from 'lib/constants';
-import { formatIsoDateConversion } from 'lib/utils';
+import { formatIsoDateConversion } from 'lib/datetime';
 import { MKRWeightTimeRanges } from '../delegates.constants';
 import { MKRWeightHisory } from '../types/mkrWeight';
 import getMaker from 'lib/maker';
